requirements.txt
fastapi
uvicorn
python-dotenv
openai
pillow
gunicorn
